[base] [LINUXEP-6008] Tidy handling of streams in ProcessImpl.
Add logging and move some functionality from watchdog_main into PluginProxy. (stop(IfRunning) and startIfRequired.

+++ base/modules/watchdog/watchdogimpl/PluginProxy.cpp
@@ -7,13 +7,15 @@ Copyright 2018, Sophos Limited.  All rights reserved.
 #include "PluginProxy.h"
 #include "Logger.h"
 
-watchdog::watchdogimpl::PluginProxy::PluginProxy(Common::PluginRegistryImpl::PluginInfo info)
-    : m_info(std::move(info)),m_process(Common::Process::createProcess())
+using namespace watchdog::watchdogimpl;
+
+PluginProxy::PluginProxy(Common::PluginRegistryImpl::PluginInfo info)
+    : m_info(std::move(info)),m_process(Common::Process::createProcess()),m_running(false),m_deathTime(0)
 {
     m_exe = m_info.getExecutableFullPath();
 }
 
-void watchdog::watchdogimpl::PluginProxy::start()
+void PluginProxy::start()
 {
     if (m_exe.empty())
     {
@@ -25,19 +27,23 @@ void watchdog::watchdogimpl::PluginProxy::start()
         m_info.getExecutableArguments(),
         m_info.getExecutableEnvironmentVariables()
         );
+    m_running = true;
 }
 
-void watchdog::watchdogimpl::PluginProxy::stop()
+void PluginProxy::stop()
 {
     if (m_exe.empty())
     {
         return;
     }
-    LOGINFO("Stopping "<<m_exe);
-    m_process->kill();
+    if (m_running)
+    {
+        LOGINFO("Stopping " << m_exe);
+        m_process->kill();
+    }
 }
 
-Common::Process::ProcessStatus watchdog::watchdogimpl::PluginProxy::status()
+Common::Process::ProcessStatus PluginProxy::status()
 {
     if (m_exe.empty())
     {
@@ -46,7 +52,7 @@ Common::Process::ProcessStatus watchdog::watchdogimpl::PluginProxy::status()
     return m_process->getStatus();
 }
 
-int watchdog::watchdogimpl::PluginProxy::exitCode()
+int PluginProxy::exitCode()
 {
     if (m_exe.empty())
     {
@@ -56,5 +62,47 @@ int watchdog::watchdogimpl::PluginProxy::exitCode()
 
     LOGINFO(m_exe<<" exited with "<<code);
 
+    std::string output = m_process->output();
+    LOGINFO("Output: "<<output);
+
     return code;
+}
+
+void PluginProxy::checkForExit()
+{
+    if (!m_running)
+    {
+        // Don't print out multiple times
+        return;
+    }
+    if (status() == Common::Process::ProcessStatus::FINISHED)
+    {
+        int code = exitCode();
+        if (code != 0)
+        {
+            LOGERROR("Process died with " << code);
+        }
+        m_running = false;
+        m_deathTime = ::time(nullptr);
+    }
+}
+
+time_t PluginProxy::startIfRequired()
+{
+    if (m_running)
+    {
+        return 3600;
+    }
+
+    time_t now = ::time(nullptr);
+    if ((now - m_deathTime) > 10)
+    {
+        start();
+        return 3600;
+    }
+    else
+    {
+        LOGDEBUG("Not starting "<<m_exe);
+        return 10 - (now - m_deathTime);
+    }
 }
